requirements.txt
Flask==2.3.2
pyswisseph==2.10.02
